worker/text-preparation-worker/requirements.txt
beautifulsoup4==4.5.3
redis
pypdf2
Pillow
docx
lxml
pytesseract